% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_dalle.R
\name{geo_dalle}
\alias{geo_dalle}
\title{Raster du potentiel de réserve profondes}
\usage{
geo_dalle(
  mnt0,
  zn_prf = NULL,
  zn = NULL,
  geo_meuble = NULL,
  p.pendage = NULL,
  con_geo = NULL,
  db_geo = NULL,
  db_pendage = NULL,
  geo_not_dalle = c("j", "n", "c", "l"),
  buff = 500
)
}
\arguments{
\item{mnt0}{raster MNT, à résolution minimale de 5m}

\item{zn_prf}{polygone (sf) de la zone à étudier}

\item{zn}{polygone sf de découpage du MNT}

\item{geo_meuble}{polygones des roches meubles (moraines, molasses, éboulis, alluvions...)}

\item{p.pendage}{points d'observation BRGM des pendages}

\item{con_geo}{connexion database}

\item{db_geo}{nom de la table géologie}

\item{db_pendage}{nom de la table des points de pendafe}

\item{geo_not_dalle}{vecteur des premières lettres des notations des couches de roches non meubles}

\item{buff}{buffer (en m) à utiliser autour de zn_prf pour réaliser les calculs (défaut: 500m)}
}
\value{
raster du risque de présence d'une dalle compact
}
\description{
Raster du potentiel de réserve profondes
}
\examples{
dir <- system.file(package = "geoDalle")
mnt <- raster(file.path(dir, "data","mnt0.tif"))
zn_prf <- read_sf(file.path(dir,"data","zn_prf.shp"))
geo_meuble <- read_sf(file.path(dir,"data","geo_meuble.shp"))
p.pendage <- read_sf(file.path(dir,"data","p.pendage.shp"))
geo_dalle(mnt0=mnt, zn_prf=zn_prf, geo_meuble=geo_meuble, p.pendage=p.pendage)

}
