% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.R
\name{qm}
\alias{qm}
\title{Carte d'un raster}
\usage{
qm(r, n = NULL, pal = "Spectral", title = NULL, breaks = NULL)
}
\arguments{
\item{r}{raster}

\item{n}{nb de couleurs}

\item{pal}{palette de couleurs}

\item{title}{titre}

\item{breaks}{vecteur des limites de couleurs}
}
\value{
carte tmap
}
\description{
Carte d'un raster
}
