% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_dalle.R
\name{geo_dalle}
\alias{geo_dalle}
\title{Raster des risques de dalles}
\usage{
geo_dalle(
  mnt0,
  zn_prf = NULL,
  zn = NULL,
  geo_meuble = NULL,
  p.pendage = NULL,
  con_geo = NULL,
  db_geo = NULL,
  db_pendage = NULL,
  geo_not_dalle = c("j", "n", "c", "l")
)
}
\arguments{
\item{mnt0}{raster MNT, à résolution minimale de 5m}

\item{zn_prf}{polygone (sf) de la zone à étudier}

\item{zn}{polygone sf de découpage du MNT}

\item{geo_meuble}{polygones des roches meubles (moraines, molasses, éboulis, alluvions...)}

\item{p.pendage}{points d'observation BRGM des pendages}

\item{con_geo}{connexion database}

\item{db_geo}{nom de la table géologie}

\item{db_pendage}{nom de la table des points de pendafe}

\item{geo_not_dalle}{vecteur des premières lettres des notations des couches de roches non meubles}
}
\value{
raster du risque de présence d'une dalle compact
}
\description{
A croiser avec la nature géologique, afin d'éliminer les formations
meubles (alluvions, moraines...)
}
